Split utilities for compiler output handling into a separate header
